+++ cmd/audius-ctl/register.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/AudiusProject/audius-d/pkg/conf"
 	"github.com/AudiusProject/audius-d/pkg/logger"
+	"github.com/AudiusProject/audius-d/pkg/orchestration"
 	"github.com/AudiusProject/audius-d/pkg/register"
 	"github.com/spf13/cobra"
 )
@@ -15,20 +16,24 @@ var registerCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx_config, err := conf.ReadOrCreateContextConfig()
 		if err != nil {
-			return logger.Error("Failed to retrieve context: ", err)
+			return logger.Error("Failed to retrieve context:", err)
 		}
 		for host, nodeConfig := range ctx_config.Nodes {
 			if nodeConfig.Type != conf.Creator {
 				continue
 			}
-			err := register.RegisterNode(
+			privateKey, err := orchestration.NormalizedPrivateKey(host, nodeConfig.PrivateKey)
+			if err != nil {
+				return logger.Error("Failed to obtain private key:", err)
+			}
+			err = register.RegisterNode(
 				"content-node",
 				fmt.Sprintf("https://%s", host),
 				"http://localhost:8546",
 				register.GanacheAudiusTokenAddress,
 				register.GanacheContractRegistryAddress,
 				nodeConfig.Wallet,
-				nodeConfig.PrivateKey,
+				privateKey,
 			)
 			if err != nil {
 				return logger.Error("Failed to register node:", err)
