+++ pkg/orchestration/run.go
@@ -1,10 +1,13 @@
 package orchestration
 
 import (
+	"bytes"
 	"fmt"
 	"io"
+	"net"
 	"os"
 	"os/exec"
+	"strings"
 
 	"github.com/AudiusProject/audius-d/pkg/conf"
 	"github.com/AudiusProject/audius-d/pkg/logger"
@@ -72,6 +75,24 @@ func RunDownNodes(nodes map[string]conf.NodeConfig) {
 	}
 }
 
+func NormalizedPrivateKey(host, privateKeyConfigValue string) (string, error) {
+	privateKey := privateKeyConfigValue
+	if strings.HasPrefix(privateKeyConfigValue, "/") {
+		// get key value from file on host
+		outBuf := new(bytes.Buffer)
+		errBuf := new(bytes.Buffer)
+		if err := execOnHost(host, outBuf, errBuf, "cat", privateKeyConfigValue); err != nil {
+			return "", logger.Error(errBuf.String(), err)
+		}
+		privateKey = strings.TrimSpace(outBuf.String())
+	}
+	privateKey = strings.TrimPrefix(privateKey, "0x")
+	if len(privateKey) != 64 {
+		return "", logger.Error("Invalid private key")
+	}
+	return privateKey, nil
+}
+
 func execLocal(command string, args ...string) error {
 	cmd := exec.Command(command, args...)
 	cmd.Stdout = os.Stdout
@@ -79,9 +100,31 @@ func execLocal(command string, args ...string) error {
 	return cmd.Run()
 }
 
-func execRemote(host string, stdout io.Writer, stderr io.Writer, command string, args ...string) error {
-	cmd := exec.Command("ssh", append([]string{host, command}, args...)...)
+func execOnHost(host string, stdout io.Writer, stderr io.Writer, command string, args ...string) error {
+	var cmd *exec.Cmd
+	isLocalhost, err := resolvesToLocalhost(host)
+	if err != nil {
+		return logger.Error("Error determining origin of host:", err)
+	} else if isLocalhost {
+		cmd = exec.Command(command, args...)
+	} else {
+		cmd = exec.Command("ssh", append([]string{host, command}, args...)...)
+	}
 	cmd.Stdout = stdout
 	cmd.Stderr = stderr
 	return cmd.Run()
+}
+
+func resolvesToLocalhost(host string) (bool, error) {
+	ips, err := net.LookupHost(host)
+	if err != nil {
+		return false, logger.Errorf("Cannot resolve host %s: %s", host, err.Error())
+	}
+
+	for _, ip := range ips {
+		if ip == "127.0.0.1" || ip == "::1" {
+			return true, nil
+		}
+	}
+	return false, nil
 }
